requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
python-dotenv
passlib[bcrypt]
bcrypt==4.0.1
python-jose
pydantic[email]
python-multipart
python-jose[cryptography]
reportlab
matplotlib
fastapi-mail
firebase-admin
ollama
groq
apscheduler
gunicorn
postgresql
alembic
httpx